--- src/components/pages/MyMentees/data.json
@@ -1,237 +1,0 @@
-[
-  {
-    "profile_id": "601d51ed-6016-4fa9-ae49-fe50b082e8c3",
-    "first_name": "Kingston",
-    "last_name": "Morris",
-    "email": "Kingston.Morris@gmail.com",
-    "country": "U.S.",
-    "state": "Missouri",
-    "city": "Rochester",
-    "current_company": "Microsoft",
-    "current_position": "Data Engineer",
-    "tech_stack": ["Backend", "Design UI/UX", "Career Development"],
-    "commitment": true,
-    "job_help": false,
-    "industry_knowledge": false,
-    "pair_programming": true,
-    "role" : "mentee",
-    "referred_by": "Meta",
-    "other_info": "anything else may be written here",
-    "validate_status": "pending",
-    "is_active": true,
-    "accepting_new_mentees": true,
-    "created_at": "2022-10-31T19:36:48.231000"
-  },
-  {
-    "profile_id": "fce3b89c-4d8e-4547-bdfc-2ffa6e7b1492",
-    "first_name": "Rhys",
-    "last_name": "Parker",
-    "email": "Rhys.Parker@gmail.com",
-    "country": "U.S.",
-    "state": "Maryland",
-    "role" : "mentee",
-    "city": "New Town",
-    "current_company": "Meta",
-    "current_position": "Project Manager",
-    "tech_stack": ["Data Science", "iOS", "Android"],
-    "commitment": true,
-    "job_help": false,
-    "industry_knowledge": false,
-    "pair_programming": false,
-    "referred_by": "Meta",
-    "other_info": "anything else may be written here",
-    "validate_status": "pending",
-    "is_active": true,
-    "accepting_new_mentees": true,
-    "created_at": "2022-10-31T19:36:48.231000"
-  },
-  {
-    "profile_id": "f299125d-15a7-4721-a3b1-498733fa5e02",
-    "first_name": "Adonis",
-    "last_name": "Taylor",
-    "email": "Adonis.Taylor@gmail.com",
-    "country": "U.S.",
-    "state": "Alabama",
-    "role" : "mentor",
-    "city": "Washington",
-    "current_company": "Acme",
-    "current_position": "Data Scientist",
-    "tech_stack": ["Career Development"],
-    "commitment": true,
-    "job_help": false,
-    "industry_knowledge": false,
-    "pair_programming": false,
-    "referred_by": "Meta",
-    "other_info": "anything else may be written here",
-    "validate_status": "pending",
-    "is_active": true,
-    "accepting_new_mentees": true,
-    "created_at": "2022-10-31T19:36:48.231000"
-  },
-  {
-    "profile_id": "601d51ed-6016-4fa9-ae49-fe50b082e8c3",
-    "first_name": "George",
-    "last_name": "Kelly",
-    "email": "George.Kelly@gmail.com",
-    "country": "U.S.",
-    "state": "Missouri",
-    "city": "Rochester",
-    "role" : "mentor",
-    "current_company": "Meta",
-    "current_position": "Data Engineer",
-    "tech_stack": [
-      "Data Science",
-      "Android",
-      "Design UI/UX",
-      "Career Development"
-    ],
-    "commitment": true,
-    "job_help": false,
-    "industry_knowledge": true,
-    "pair_programming": false,
-    "referred_by": "Colleague",
-    "other_info": "anything else may be written here",
-    "validate_status": "pending",
-    "is_active": true,
-    "accepting_new_mentees": true,
-    "created_at": "2022-10-31T19:36:48.231000"
-  },
-  {
-    "profile_id": "57c73735-def9-4d2d-9587-617e246422e8",
-    "first_name": "Raiden",
-    "last_name": "James",
-    "email": "Raiden.James@gmail.com",
-    "country": "U.S.",
-    "state": "Rhode Island",
-    "city": "Los Alamos",
-    "role" : "mentor",
-    "current_company": "Netflix",
-    "current_position": "Software Engineer",
-    "tech_stack": ["Data Science"],
-    "commitment": true,
-    "job_help": true,
-    "industry_knowledge": false,
-    "pair_programming": true,
-    "referred_by": "YouTube",
-    "other_info": "anything else may be written here",
-    "validate_status": "pending",
-    "is_active": true,
-    "accepting_new_mentees": false,
-    "created_at": "2022-10-31T19:36:48.231000"
-  },
-  {
-    "profile_id": "7e8b7ff4-eb28-4688-b7ee-4c5008db11c7",
-    "first_name": "Harley",
-    "last_name": "Garcia",
-    "email": "Harley.Garcia@gmail.com",
-    "country": "U.S.",
-    "state": "North Dakota",
-    "city": "Rochester",
-    "current_company": "Alphabet",
-    "role" : "mentee",
-    "current_position": "Project Manager",
-    "tech_stack": ["Backend", "Design UI/UX"],
-    "commitment": true,
-    "job_help": false,
-    "industry_knowledge": false,
-    "pair_programming": false,
-    "referred_by": "Colleague",
-    "other_info": "anything else may be written here",
-    "validate_status": "pending",
-    "is_active": true,
-    "accepting_new_mentees": true,
-    "created_at": "2022-10-31T19:36:48.231000"
-  },
-  {
-    "profile_id": "601d51ed-6016-4fa9-ae49-fe50b082e8c3",
-    "first_name": "Winston",
-    "last_name": "Reyes",
-    "email": "Winston.Reyes@gmail.com",
-    "country": "U.S.",
-    "state": "Virgin Islands",
-    "city": "Los Alamos",
-    "current_company": "Oracle",
-    "current_position": "Full Stack Engineer",
-    "tech_stack": ["Data Science", "iOS", "Backend"],
-    "commitment": true,
-    "job_help": false,
-    "industry_knowledge": false,
-    "role" : "mentee",
-    "pair_programming": false,
-    "referred_by": "Friend",
-    "other_info": "anything else may be written here",
-    "validate_status": "pending",
-    "is_active": false,
-    "accepting_new_mentees": true,
-    "created_at": "2022-10-31T19:36:48.231000"
-  },
-  {
-    "profile_id": "ac985dd8-9164-4825-833b-aef12264db8f",
-    "first_name": "Janiyah",
-    "last_name": "Morris",
-    "email": "Janiyah.Morris@gmail.com",
-    "country": "U.S.",
-    "state": "District ",
-    "city": "Los Alamos",
-    "formerly_incarcerated": false,
-    "underrepresented_group": false,
-    "low_income": true,
-    "convictions": "Misdemeanor",
-    "tech_stack": "Backend",
-    "role" : "mentee",
-    "job_help": true,
-    "pair_programming": false,
-    "referred_by": "Meta",
-    "other_info": "anything else may be written here",
-    "validate_status": "pending",
-    "is_active": true,
-    "in_project_underdog": true,
-    "created_at": "2022-10-31T14:10:22.316000"
-  },
-  {
-    "profile_id": "c6af50a3-797a-4e09-b3b2-2cbf0b2baea3",
-    "first_name": "Kade",
-    "last_name": "Carter",
-    "email": "Kade.Carter@gmail.com",
-    "country": "U.S.",
-    "state": "Indiana",
-    "city": "New Town",
-    "formerly_incarcerated": false,
-    "underrepresented_group": true,
-    "low_income": false,
-    "convictions": "Felony",
-    "tech_stack": "Career Development",
-    "job_help": true,
-    "role" : "mentor",
-    "pair_programming": false,
-    "referred_by": "Colleague",
-    "other_info": "anything else may be written here",
-    "validate_status": "pending",
-    "is_active": true,
-    "in_project_underdog": false,
-    "created_at": "2022-10-31T14:10:22.315000"
-  },
-  {
-    "profile_id": "f6ebc220-bd97-4ef7-a364-8d4a016fe38a",
-    "first_name": "Zachariah",
-    "last_name": "Kelly",
-    "email": "Zachariah.Kelly@gmail.com",
-    "country": "U.S.",
-    "state": "of Columbia",
-    "role" : "mentor",
-    "city": "Arrowhead",
-    "formerly_incarcerated": true,
-    "underrepresented_group": false,
-    "low_income": true,
-    "convictions": "Misdemeanor",
-    "tech_stack": "Android",
-    "job_help": false,
-    "pair_programming": false,
-    "referred_by": "Twitter",
-    "other_info": "anything else may be written here",
-    "validate_status": "approved",
-    "is_active": true,
-    "in_project_underdog": false,
-    "created_at": "2022-10-31T14:10:22.315000"
-  }
-]
