+++ src/index.js
@@ -13,9 +13,9 @@ import { Landing } from './components/pages/LandingPage';
 import Apply from './components/pages/RoleApply/Apply';
 import Mentee from './components/pages/RoleApply/Applications/Mentee';
 import Mentor from './components/pages/RoleApply/Applications/Mentor';
-import MyMentors from './components/pages/MyMentors/MyMentors';
+import MyMentors from './components/pages/Mentee-MentorDashboard/MyMentors';
 import MyMemos from './components/pages/Memos/MyMemos';
-import MyMentees from './components/pages/MyMentees/MyMentees';
+import MyMentees from './components/pages/Mentee-MentorDashboard/MentorDashboard/Mentee-MentorDashboard';
 //import ViewAllMeetings from './components/pages/ViewAllMeetings/ViewAllMeetings'; this component has been changed and now exists in 3 places
 import AdminMeetingDash from './components/pages/AdminMeetingsDashBoard/AdminMeetingDash';
 import MentorMeetingDash from './components/pages/MentorMeetingsDashBoard/MentorMeetingDash';
@@ -107,6 +107,7 @@ function App() {
           allowRoles={[1, 2, 5]}
           component={UserManagement}
         />
+
         <PrivateRoute
           path="/mentees"
           allowRoles={[1, 2, 3, 5]}
